Added telemetry, DB snapshots, fixed bugs, added some stuff
Replaced file snapshots with snapshots to either SQLite
or Postgres, with a new server `Name` config to differentiate
snapshots from different instances.

Added telemetry, and updated the compose file to include a
jaeger instance to view traces. Metrics, expvar and
pprof endpoints are now available on a separate port
(`33970` by default) as `/metrics`, `/debug/pprof` and
`/debug/vars`. These are only enabled via configuration
individually (`KEYQUARRY_METRICS=true`, `KEYQUARRY_PPROF=true`
and/or `KEYQUARRY_EXPVAR=true` envvars, for each).

Renamed `GetKeyInfo` to `Inspect` and added an option to
return the value of the key along with metadata.

Added a stream `WatchKeyValue`, which subscribes the client
to the key, immediately sending the current value of the key
(with some metadata), and sending updated data whenever the
content of the key changes.

Metrics now return "key pressure" (how close the server is
to the key limit)

Added key-specific metrics, which persist across multiple
instances of the same key. So if a key is deleted/expired/expunged,
this will indicate the first time a key with that name was created,
not just when the current instance was created.

Fixed some potential deadlock scenarios

Fixed an issue where an event subscriber would sometimes only
be unsubscribed if they attempted to re-subscribe.

Snapshots now take into account lifespans, and key-specific metrics,
which are reloaded with the snapshot.

Value hashing is now just done via `hash/fnv` for speed and simplicity.

Lifespans now override locks.

Added a hard stop to the GRPC server if a graceful stop takes too long,
which is configurable.

Added/removed/updated a bunch of tests